backend/requirements.txt
langchain
langchain-community
python-dotenv
pydantic
fastapi
pypdf
google-generativeai
uvicorn[standard]
pipdeptree
Markdown